// src/types.ts
export interface Car {
  id: number;
  make: string;
  model: string;
  version: string;
  price: number;
  year: number;
  km: number;
  fuel: string;
  power: number;
  country: string;
  city: string;
  source: string;
  img: string;
}

export interface Country {
  name: string;
  flag: string;
}